Split objectclass out of object.c/.h
