+++ src/builtin/veneer.h
@@ -11,6 +11,8 @@
 
 #include "upvalue.h"
 #include "function.h"
+#include "clss.h"
+#include "cmplx.h"
 #include "closure.h"
 #include "invocation.h"
 #include "instance.h"
@@ -20,15 +22,14 @@
 #include "strng.h"
 #include "dict.h"
 #include "err.h"
+//#include "file.h"
+//#include "system.h"
 
 #include "matrix.h"
 
 /* ---------------------------
  * Veneer classes
  * --------------------------- */
-
-#define CLASS_INVK                        "ClssInvk"
-#define CLASS_INVK_MSG                    "Cannot invoke method '%s' on a class."
  
 void veneer_initialize(void);
